+++ app/src/main/java/cn/wthee/pcrtool/ui/character/CharacterSkill.kt
@@ -5,7 +5,7 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import cn.wthee.pcrtool.ui.compose.CommonSpacer
+import cn.wthee.pcrtool.ui.common.CommonSpacer
 import cn.wthee.pcrtool.ui.skill.SkillCompose
 import cn.wthee.pcrtool.ui.theme.Dimen
 import coil.annotation.ExperimentalCoilApi
