+++ app/src/main/java/cn/wthee/pcrtool/ui/character/CharacterStatusCoe.kt
@@ -14,7 +14,7 @@ import cn.wthee.pcrtool.R
 import cn.wthee.pcrtool.data.db.entity.UnitStatusCoefficient
 import cn.wthee.pcrtool.data.model.AttrValue
 import cn.wthee.pcrtool.ui.PreviewBox
-import cn.wthee.pcrtool.ui.compose.*
+import cn.wthee.pcrtool.ui.common.*
 import cn.wthee.pcrtool.ui.theme.Dimen
 import cn.wthee.pcrtool.utils.Constants
 import cn.wthee.pcrtool.viewmodel.CharacterAttrViewModel
