+++ app/src/main/java/cn/wthee/pcrtool/ui/character/RankEquipList.kt
@@ -20,7 +20,7 @@ import cn.wthee.pcrtool.R
 import cn.wthee.pcrtool.data.db.entity.UnitPromotion
 import cn.wthee.pcrtool.ui.MainActivity.Companion.navViewModel
 import cn.wthee.pcrtool.ui.PreviewBox
-import cn.wthee.pcrtool.ui.compose.*
+import cn.wthee.pcrtool.ui.common.*
 import cn.wthee.pcrtool.ui.theme.Dimen
 import cn.wthee.pcrtool.viewmodel.EquipmentViewModel
 import coil.annotation.ExperimentalCoilApi
