+++ app/src/main/java/cn/wthee/pcrtool/ui/tool/TweetList.kt
@@ -29,7 +29,7 @@ import cn.wthee.pcrtool.data.db.entity.urlGetId
 import cn.wthee.pcrtool.data.enums.MainIconType
 import cn.wthee.pcrtool.data.model.TweetButtonData
 import cn.wthee.pcrtool.ui.PreviewBox
-import cn.wthee.pcrtool.ui.compose.*
+import cn.wthee.pcrtool.ui.common.*
 import cn.wthee.pcrtool.ui.theme.Dimen
 import cn.wthee.pcrtool.utils.Constants
 import cn.wthee.pcrtool.utils.openWebView
