+++ app/src/main/java/cn/wthee/pcrtool/ui/tool/pvp/PvpFloatSearch.kt
@@ -25,7 +25,7 @@ import cn.wthee.pcrtool.R
 import cn.wthee.pcrtool.data.enums.MainIconType
 import cn.wthee.pcrtool.ui.FabMain
 import cn.wthee.pcrtool.ui.MainActivity
-import cn.wthee.pcrtool.ui.compose.FabCompose
+import cn.wthee.pcrtool.ui.common.FabCompose
 import cn.wthee.pcrtool.ui.theme.Dimen
 import cn.wthee.pcrtool.ui.theme.PcrtoolComposeTheme
 import cn.wthee.pcrtool.utils.VibrateUtil
