+++ app/src/main/java/cn/wthee/pcrtool/utils/ScreenUtil.kt
@@ -4,7 +4,7 @@ import android.os.Build
 import android.util.DisplayMetrics
 import android.view.View
 import cn.wthee.pcrtool.MyApplication
-import cn.wthee.pcrtool.ui.compose.RATIO
+import cn.wthee.pcrtool.ui.common.RATIO
 
 
 /**
